requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
beanie==1.25.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
lazy-model==0.2.0
mongomock==4.1.2
mongomock_motor==0.0.29
motor==3.4.0
packaging==24.0
passlib==1.7.4
pluggy==1.4.0
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pymongo==4.6.3
pytest==8.1.1
pytest-mock==3.14.0
sentinels==1.0.0
sniffio==1.3.1
starlette==0.37.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
pydantic_settings
pandas
scikit-learn==1.6.0